requirements.txt
APScheduler==3.6.3
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
protobuf==3.17.2
PyMySQL==1.0.2
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.17
tornado==6.1
tzlocal==2.1
urllib3==1.26.5
Werkzeug==2.0.1
